# feat: MCP server implementation with hot-reload development improvements