--- a/packages/cli/vite.config.ts
+++ b/packages/cli/vite.config.ts
@@ -27,6 +27,7 @@ export default defineConfig({
         'node:path',
         'node:os',
         'node:url',
+        'node:child_process',
       ],
       output: {
         banner: chunk => {