--- a/packages/core/src/application/usecases/complete-ticket.ts
+++ b/packages/core/src/application/usecases/complete-ticket.ts
@@ -1,26 +1,18 @@
 import { ERROR_MESSAGES, TicketNotFoundError } from '@project-manager/shared'
-import { inject, injectable } from 'inversify'
 import { TicketId } from '../../domain/value-objects/ticket-id.js'
 import type { UseCase } from '../common/base-usecase.js'
 import { CompleteTicketRequest } from '../dtos/requests/complete-ticket.js'
 import { CompleteTicketResponse } from '../dtos/responses/complete-ticket.js'
 import { TicketResponse } from '../dtos/responses/ticket.js'
-import {
-  type TicketRepository,
-  TicketRepository as TicketRepositorySymbol,
-} from '../repositories/ticket-repository.js'
+import type { TicketRepository } from '../repositories/ticket-repository.js'
 
 /**
  * Use case for completing a ticket.
  */
-@injectable()
 export class CompleteTicketUseCase
   implements UseCase<CompleteTicketRequest, CompleteTicketResponse>
 {
-  constructor(
-    @inject(TicketRepositorySymbol)
-    private readonly ticketRepository: TicketRepository
-  ) {}
+  constructor(private readonly ticketRepository: TicketRepository) {}
 
   async execute(request: CompleteTicketRequest): Promise<CompleteTicketResponse> {
     const ticketId = TicketId.create(request.id)