--- a/packages/core/src/application/usecases/get-ticket-by-id.ts
+++ b/packages/core/src/application/usecases/get-ticket-by-id.ts
@@ -1,26 +1,18 @@
-import { inject, injectable } from 'inversify'
 import { TicketId } from '../../domain/value-objects/ticket-id.js'
 import type { UseCase } from '../common/base-usecase.js'
 import { GetTicketByIdRequest } from '../dtos/requests/get-ticket-by-id.js'
 import { GetTicketByIdResponse } from '../dtos/responses/get-ticket-by-id.js'
 import { TicketResponse } from '../dtos/responses/ticket.js'
-import {
-  type TicketRepository,
-  TicketRepository as TicketRepositorySymbol,
-} from '../repositories/ticket-repository.js'
+import type { TicketRepository } from '../repositories/ticket-repository.js'
 
 /**
  * Use case for retrieving a ticket by its ID.
  * Returns null if ticket is not found.
  */
-@injectable()
 export class GetTicketByIdUseCase
   implements UseCase<GetTicketByIdRequest, GetTicketByIdResponse | null>
 {
-  constructor(
-    @inject(TicketRepositorySymbol)
-    private readonly ticketRepository: TicketRepository
-  ) {}
+  constructor(private readonly ticketRepository: TicketRepository) {}
 
   async execute(request: GetTicketByIdRequest): Promise<GetTicketByIdResponse | null> {
     const ticketId = TicketId.create(request.id)