--- a/packages/core/src/application/usecases/update-ticket-description.ts
+++ b/packages/core/src/application/usecases/update-ticket-description.ts
@@ -1,26 +1,18 @@
 import { ERROR_MESSAGES, TicketNotFoundError } from '@project-manager/shared'
-import { inject, injectable } from 'inversify'
 import { TicketId } from '../../domain/value-objects/ticket-id.js'
 import type { UseCase } from '../common/base-usecase.js'
 import { UpdateTicketDescriptionRequest } from '../dtos/requests/update-ticket-description.js'
 import { TicketResponse } from '../dtos/responses/ticket.js'
 import { UpdateTicketDescriptionResponse } from '../dtos/responses/update-ticket-description.js'
-import {
-  type TicketRepository,
-  TicketRepository as TicketRepositorySymbol,
-} from '../repositories/ticket-repository.js'
+import type { TicketRepository } from '../repositories/ticket-repository.js'
 
 /**
  * Use case for updating a ticket's description.
  */
-@injectable()
 export class UpdateTicketDescriptionUseCase
   implements UseCase<UpdateTicketDescriptionRequest, UpdateTicketDescriptionResponse>
 {
-  constructor(
-    @inject(TicketRepositorySymbol)
-    private readonly ticketRepository: TicketRepository
-  ) {}
+  constructor(private readonly ticketRepository: TicketRepository) {}
 
   async execute(request: UpdateTicketDescriptionRequest): Promise<UpdateTicketDescriptionResponse> {
     const ticketId = TicketId.create(request.id)