--- a/packages/core/src/application/usecases/update-ticket-priority.ts
+++ b/packages/core/src/application/usecases/update-ticket-priority.ts
@@ -1,26 +1,18 @@
 import { ERROR_MESSAGES, TicketNotFoundError } from '@project-manager/shared'
-import { inject, injectable } from 'inversify'
 import { TicketId } from '../../domain/value-objects/ticket-id.js'
 import type { UseCase } from '../common/base-usecase.js'
 import { UpdateTicketPriorityRequest } from '../dtos/requests/update-ticket-priority.js'
 import { TicketResponse } from '../dtos/responses/ticket.js'
 import { UpdateTicketPriorityResponse } from '../dtos/responses/update-ticket-priority.js'
-import {
-  type TicketRepository,
-  TicketRepository as TicketRepositorySymbol,
-} from '../repositories/ticket-repository.js'
+import type { TicketRepository } from '../repositories/ticket-repository.js'
 
 /**
  * Use case for updating a ticket's priority.
  */
-@injectable()
 export class UpdateTicketPriorityUseCase
   implements UseCase<UpdateTicketPriorityRequest, UpdateTicketPriorityResponse>
 {
-  constructor(
-    @inject(TicketRepositorySymbol)
-    private readonly ticketRepository: TicketRepository
-  ) {}
+  constructor(private readonly ticketRepository: TicketRepository) {}
 
   async execute(request: UpdateTicketPriorityRequest): Promise<UpdateTicketPriorityResponse> {
     const ticketId = TicketId.create(request.id)