--- a/packages/core/src/application/usecases/update-ticket-status.ts
+++ b/packages/core/src/application/usecases/update-ticket-status.ts
@@ -1,26 +1,18 @@
 import { ERROR_MESSAGES, TicketNotFoundError } from '@project-manager/shared'
-import { inject, injectable } from 'inversify'
 import { TicketId } from '../../domain/value-objects/ticket-id.js'
 import type { UseCase } from '../common/base-usecase.js'
 import { UpdateTicketStatusRequest } from '../dtos/requests/update-ticket-status.js'
 import { TicketResponse } from '../dtos/responses/ticket.js'
 import { UpdateTicketStatusResponse } from '../dtos/responses/update-ticket-status.js'
-import {
-  type TicketRepository,
-  TicketRepository as TicketRepositorySymbol,
-} from '../repositories/ticket-repository.js'
+import type { TicketRepository } from '../repositories/ticket-repository.js'
 
 /**
  * Use case for updating a ticket's status.
  */
-@injectable()
 export class UpdateTicketStatusUseCase
   implements UseCase<UpdateTicketStatusRequest, UpdateTicketStatusResponse>
 {
-  constructor(
-    @inject(TicketRepositorySymbol)
-    private readonly ticketRepository: TicketRepository
-  ) {}
+  constructor(private readonly ticketRepository: TicketRepository) {}
 
   async execute(request: UpdateTicketStatusRequest): Promise<UpdateTicketStatusResponse> {
     const ticketId = TicketId.create(request.id)