--- a/packages/core/src/application/usecases/update-ticket-title.ts
+++ b/packages/core/src/application/usecases/update-ticket-title.ts
@@ -1,26 +1,18 @@
 import { ERROR_MESSAGES, TicketNotFoundError } from '@project-manager/shared'
-import { inject, injectable } from 'inversify'
 import { TicketId } from '../../domain/value-objects/ticket-id.js'
 import type { UseCase } from '../common/base-usecase.js'
 import { UpdateTicketTitleRequest } from '../dtos/requests/update-ticket-title.js'
 import { TicketResponse } from '../dtos/responses/ticket.js'
 import { UpdateTicketTitleResponse } from '../dtos/responses/update-ticket-title.js'
-import {
-  type TicketRepository,
-  TicketRepository as TicketRepositorySymbol,
-} from '../repositories/ticket-repository.js'
+import type { TicketRepository } from '../repositories/ticket-repository.js'
 
 /**
  * Use case for updating a ticket's title.
  */
-@injectable()
 export class UpdateTicketTitleUseCase
   implements UseCase<UpdateTicketTitleRequest, UpdateTicketTitleResponse>
 {
-  constructor(
-    @inject(TicketRepositorySymbol)
-    private readonly ticketRepository: TicketRepository
-  ) {}
+  constructor(private readonly ticketRepository: TicketRepository) {}
 
   async execute(request: UpdateTicketTitleRequest): Promise<UpdateTicketTitleResponse> {
     const ticketId = TicketId.create(request.id)