--- a/packages/mcp-server/src/utils/container.ts
+++ b/packages/mcp-server/src/utils/container.ts
@@ -1,0 +1,7 @@
+import { getContainer as getCoreContainer } from '@project-manager/core'
+import { getConfig } from '@project-manager/shared'
+
+export function getContainer() {
+  const config = getConfig()
+  return getCoreContainer(config.storagePath)
+}