… CLAUDE.md
… README.md
… docs/guides/development-tips.md
… docs/guides/dogfooding-workflow.md
… package.json
… packages/cli/package.json
… packages/cli/src/utils/config.test.ts
… packages/core/package.json
… packages/mcp-server/README.md
… packages/mcp-server/package.json
… packages/mcp-server/src/bin/mcp-server.ts
… packages/mcp-server/src/index.test.ts
… packages/mcp-server/src/index.ts
… packages/mcp-server/src/tools/create-ticket.test.ts
… packages/mcp-server/src/tools/create-ticket.ts
… packages/mcp-server/src/tools/get-project-config.ts
… packages/mcp-server/src/tools/get-project-info.ts
… packages/mcp-server/src/tools/get-ticket-by-id.ts
… packages/mcp-server/src/tools/get-ticket-stats.ts
… packages/mcp-server/src/tools/list-tickets.ts
… packages/mcp-server/src/tools/search-tickets.ts
… packages/mcp-server/src/tools/set-project-config.ts
… packages/mcp-server/src/tools/update-ticket-status.ts
… packages/mcp-server/src/utils/dev-helpers.ts
… packages/mcp-server/tsconfig.json
… packages/mcp-server/vite.config.ts
… packages/shared/src/config.ts
… pm-dev-alias.sh
… pm-mcp-http.sh
… pnpm-lock.yaml