@kamiazya 2025-07-13
## PR Review Comments Addressed

Based on the comprehensive review feedback from CodeRabbit, Copilot, and Gemini Code Assist, I've systematically addressed all actionable items. Here's a summary of the implemented changes:

### High Priority Issues Fixed
- ✅ **Verified mcp-server-dev.ts** - File exists and functions correctly with proper NODE_ENV setup

### Medium Priority Improvements  
- ✅ **Added comprehensive tests** - Created full test suite for `update-ticket-status` tool with multiple test scenarios
- ✅ **Refactored MCP tool registration** - Introduced explicit `McpTool` interface to eliminate TypeScript circular reference issues
- ✅ **Container configuration** - Maintained explicit implementation for extensibility as requested

### Low Priority Enhancements
- ✅ **Removed redundant default handling** - Cleaned up `create-ticket.ts` Zod schema default values
- ✅ **Fixed documentation wording** - Updated README.md and development-tips.md for clarity
- ✅ **Enhanced script documentation** - Added usage examples and error handling to `pm-mcp-http.sh`
- ✅ **Response formatter utility** - Created shared `response-formatter.ts` for consistent MCP responses
- ✅ **Extended CLI shutdown timeout** - Increased graceful shutdown time for better process management
- ✅ **Improved PID file handling** - Made dev-helpers.ts path resolution more robust

### Critical TypeScript Configuration Fix
- ✅ **Included test files in type checking** - Removed test file exclusions from `tsconfig.json` to prevent future CI failures and ensure consistent type safety across the entire codebase

### Technical Improvements
- **Fixed circular reference error** (TS2502) in MCP server index.ts
- **Resolved Container type mismatches** in test files
- **Added missing privacy property** to all mock responses
- **Improved mock implementation patterns** with proper Vitest typing

All changes maintain backward compatibility while significantly improving type safety, test coverage, and code quality. The TypeScript configuration change ensures all future test files will be type-checked, preventing the issues that led to CI failures.

🤖 Generated with Claude Code