@kamiazya 2025-07-13
## Summary of Recent Improvements

This PR has been significantly enhanced with comprehensive code quality improvements and architectural optimizations. Here's a summary of the major changes:

### 🔧 TypeScript Import Extensions Standardization
- **Converted ~250 import statements** from `.js` to `.ts` extensions across 96 files
- **Added `allowImportingTsExtensions: true`** to TypeScript configuration
- **Updated coding guidelines** to help AI assistants recognize the `.ts` extension approach
- **Improved development efficiency** by enabling direct TypeScript execution via tsx

### 🚀 MCP Server Architecture Optimizations
- **Optimized stateless HTTP mode** for better performance by reusing server instances
- **Created shared `McpTool` interface** to eliminate code duplication and improve type safety
- **Fixed hardcoded version strings** by dynamically importing from package.json
- **Enhanced server startup** with proper port validation (1-65535 range)

### 🛡️ Robustness and Validation Improvements
- **Enhanced boolean parsing** to accept case-insensitive values (`true/1/yes/on`)
- **Added PID validation** in development helpers to prevent NaN comparisons
- **Fixed error handlers** to exit with non-zero codes for proper process management
- **Improved configuration system** to derive valid keys from `DEFAULT_CONFIG` instead of hardcoding

### 🧪 Test Quality and Stability
- **Fixed test instability** by improving process.exit mocking
- **Resolved conditional execution** issues to prevent main function execution during test imports
- **Updated test expectations** to match actual tool count (9 tools instead of 6)
- **Fixed production build paths** to use `.js` extensions for compiled output

### 📊 Search and Filtering Enhancements
- **Implemented field-specific search** in search-tickets tool
- **Moved filtering logic to use cases** for better performance (server-side filtering)
- **Enhanced search functionality** with proper field targeting support

### 🎯 AI-Assisted Development Focus
- **Documented AI failure reduction strategies** in coding guidelines
- **Improved import resolution** to reduce AI assistant confusion
- **Enhanced development workflow** with clearer patterns and consistent conventions

## Test Results
- ✅ **All 267 tests passing**
- ✅ **Clean TypeScript compilation**
- ✅ **No linting errors**
- ✅ **Production build verification successful**

These improvements significantly enhance code quality, maintainability, and development efficiency while maintaining full backward compatibility.